Allow timezone of user to be set null
On the client side we need a way to distinguish if a user
set a timezone or not. If a user wishes to not set a timezone
the browser should deciede the timezone.

Till now the server was setting the timezone to UTC if no
timezone was set. Now if the send timezone is null
we set the timezone to null. Also if a user is requested
we send null as timezone if no timezone is configured, rather
then sending UTC.