Change choosing the source of timezone
If we choose the timezone in DateTime we did not use the
browser time as a fallback but the timezone of the user admin
(which must not necessarily be the timezone of the current user).

We now instead try to guess the the timezone of the current user
by asking the browser. The fallback after that is still the
admin user and after that we set it to UTC. In the most cases
the browser should provide a time zone.

We need to do this because in a former change, I changed the behaviour
of the server providing a fallback for the timezone of the current user.
The server now does return null if the user did not set a timezone
and wants to use the timezone of the browser.